fix(hub): break CrdtStrategy circular-dependency warning in dts build
Hoists LwwMapState/RgaState/YjsState definitions from with-commit/crdt/crdt.ts
into kernel/types.ts (crdt.ts re-exports them, all its exports now leaf-ward).
CrdtStrategy and CrdtState stay defined in kernel/types.ts (enclave
self-containment, record-codec.ts).

That hoist alone didn't clear the warning: the actual cycle closed through
kernel/vault.ts, kernel/collection.ts, and kernel/collection-config.ts, which
imported the CrdtStrategy type from with-commit/crdt/strategy.js instead of
directly from kernel/types.js where it's defined (record-codec.ts already used
the direct form). Redirecting those three type imports removes the mutual
chunk dependency between types.ts and strategy.ts that rollup's dts bundler
was warning about.

pnpm --filter @noy-db/hub build no longer emits any circular/cycle warning;
full test suite (492/492 files, 3996/3996 tests), typecheck, lint, and
check:architecture all pass. Public CRDT type surface unchanged (verified via
root-barrel-surface-golden.test.ts and kernel-api-surface-golden.test.ts).

Closes #667
